Create import_audio.py
Python's in-built wave library is used to import the good_morning.wav audio file. Then we'll see what it looks like in byte form using the built-in readframes() method. The audio file is only a few seconds long but at 48 kHz, that means it contains 48,000 pieces of information per second.